migrate-update-weights-mixin-move: Move weight-update RPC handlers to SchedulerWeightUpdaterManager
Mechanical cut + paste for the ``migrate-update-weights-mixin`` mech move.

Cut the @staticmethods plus the module-level helpers
(``_export_static_state`` / ``_import_static_state``) from
``scheduler_update_weights_mixin.py`` and paste them into the existing
``scheduler_components/weight_updater.py`` (methods into
``SchedulerWeightUpdaterManager`` class body, helpers at module level
after the class). Supporting module-level imports (``logging`` /
``traceback`` / ``torch`` / ``Tuple`` / constants / io_struct symbols)
relocate alongside.

The source file is deleted; ``SchedulerUpdateWeightsMixin`` is dropped
from the Scheduler inheritance list and its import is removed.

Method bodies otherwise byte-identical. ``@staticmethod`` decorators
dropped; ``self: "SchedulerWeightUpdaterManager"`` annotation simplified
to bare ``self``.

The prep-form dispatch lambdas collapse to direct refs (pure prefix
transformation):
  ``lambda req: self.<method>(self.weight_updater, req)`` →
  ``self.weight_updater.<method>``.

Refactor chain ID: migrate-update-weights-mixin-move
